+++ CuiGuangfan/src/com/xingkong/ThreeSumClosest.java
@@ -1,3 +1,4 @@
+package com.xingkong;
 
 public class ThreeSumClosest {
 	public int threeSumClosest(int[] nums, int target) {
