+++ CuiGuangfan/src/com/xingkong/TrappingRainWater.java
@@ -1,3 +1,4 @@
+package com.xingkong;
 
 public class TrappingRainWater {
 	public int trap(int[] height) {
