+++ CuiGuangfan/src/com/xingkong/WordSearch.java
@@ -1,3 +1,4 @@
+package com.xingkong;
 
 public class WordSearch {
 	public boolean exist(char[][] board, String word) {
